+++ src/feature/votes/votes.module.ts
@@ -2,15 +2,14 @@ import { Module } from "@nestjs/common";
 import { VotesController } from "./votes.controller";
 import { VotesService } from "./votes.service";
 import { VoteModule } from "./vote/vote.module";
-import { CommentsModule } from "./comments/comments.module";
-
+import { CommentsModule } from "./vote/comments/comments.module";
 
 /**
  * Vote module
  */
 @Module({
-    imports: [ VoteModule, CommentsModule ],
-    controllers: [ VotesController ],
-    providers: [ VotesService ]
+  imports: [VoteModule, CommentsModule],
+  controllers: [VotesController],
+  providers: [VotesService],
 })
 export class VotesModule {}
